teams/channel-extract-messages/requirements.txt
python-dotenv
requests
pandas
openpyxl
ijson
